requirements.txt
ttkbootstrap>=1.0.0
Pillow>=9.0.0
numpy>=1.20.0
ujson>=5.0.0 